Improve Account addition and editing UI
This commit changes the account UI in two main ways:
- The maximum field length was doubled from 32 to 64.
- All regular keyboard operations are now allowed on the username
field, and all operations other than copying are allowed on the
password field.

In the process of making these changes I've refactored GuiAddAccount
and GuiEditAccount, moving most of their functionality to the base
class AbstractAccountGui, and removed the manual handling of text
field contents.  Additionally a GuiPasswordField was introduced in
order to easily manage censoring out the password and prevent copy
operations.