fix: Also fetch and assign profile properties
The singleplayer server waits for the profile to quit in order to shut down.
This profile is not currently set by IAS, but is set by MC when booting up the game
with a Mojang account, making the singleplayer server wait for a player that no longer exists.

Setting it properly (according to vanilla rules) will allow the singleplayer server to shut down.

This field only exists in 1.20.2 and newer.
See #188 and #207.