feat: Prepare to have legacy/stonecutter separation
This is a "combined" migrational project structure to
use the Stonecutter preprocessor.

Unlike the separate stonecutter branch, this will not get "desynched"
from the main branch or become obsolete after one branch is maintained
and the other is not. The plan for migration is this:

1. Move out legacy code to a separate folder, but not to a separate
branch, so the migration can be smooth and gradual. (phasing out)
2. Prepare the Stonecutter environment.
3. Slowly migrate versions to Stonecutter, phasing out the "legacy" folder.
4. Done!

At any time, the general rules of the main branch apply:

- The mod must compile. (i.e. don't push uncompilable changes)
- The mod must run. (side note: modern LexForge versions don't run on ArchLoom,
they might be tested in a separate game environment)
- All existing code must NOT be commented out to comply with two previous requirements.
- Legacy and stonecutter code versions should have the same set of features. (i.e.
no new big features and bugfixes must be done on both legacy and SC codepaths)